fix: Enable manual input of date range for super admin
